B3a Write a Java method to convert times between eastern time (ET), mountain time (MT), and coordinated universal time (UTC) zones.
Created files: TimezoneConversion.java and TimezoneConversionController.java

Declared and initialized DateTimeFormatter with HH:mm

Declared the variable currentTime for parsing and converting times into ET, MT, and UTC.

Used the formatter to convert the variables into strings and returned the converted times as a string.

Defined convertTimezone method to return the converted times from TimezoneConversion.java as a string.